backend/requirements.txt
# FastAPI and server (CORS is built-in)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.13
Pillow==10.1.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# PDF generation
fpdf2==2.7.6

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Type hints
pydantic==2.5.0